findmarket - django/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
gunicorn==20.1.0
mysqlclient==2.1.1
sqlparse==0.4.3
whitenoise==6.2.0
